Add tests for upload
